We have observed additional overhead in the `JiebaTokenizer` wrapper.

![screenshot-20250529-134726](https://github.com/user-attachments/assets/daed51a3-0fde-4603-a0de-f6b96c41591a)

It was caused by memory allocation of `indices` and `Token::text`. This PR eliminates this overhead by moving the `Token` construction into the `advance` function.

After the optimization, we added a benchmark on 围城.txt. It shows performance improvement on my Macbook M3 Pro:
```
# main
jieba_tokenizer/tokenize
                        time:   [41.821 ms 42.125 ms 42.432 ms]
                        thrpt:  [16.495 MiB/s 16.615 MiB/s 16.736 MiB/s]
# this PR
jieba_tokenizer/tokenize
                        time:   [33.987 ms 34.270 ms 34.576 ms]
                        thrpt:  [20.243 MiB/s 20.424 MiB/s 20.594 MiB/s]
                 change:
                        time:   [-19.599% -18.647% -17.670%] (p = 0.00 < 0.05)
                        thrpt:  [+21.462% +22.921% +24.377%]
                        Performance has improved.
```